Rename tests
The old names were from when cclyzer++ was part of the MATE source tree; since
cclyzer++ is primarily a pointer analysis it is redundant to say so in the test
file names.
